pyproject.toml
[project]
name = "sophie-sales-management-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic==1.16.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "bcrypt==4.3.0",
    "bs4>=0.0.2",
    "cffi==1.17.1",
    "click==8.1.8",
    "colorthief>=0.2.1",
    "cryptography==45.0.3",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.115.12",
    "greenlet==3.2.2",
    "h11==0.16.0",
    "httpx>=0.28.1",
    "idna==3.10",
    "jinja2>=3.1.6",
    "lxml>=6.0.0",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "minio>=7.2.16",
    "openai>=1.93.3",
    "passlib==1.7.4",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pyotp>=2.9.0",
    "python-dotenv==1.1.0",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "qrcode[pil]>=8.2",
    "reportlab>=4.4.2",
    "requests>=2.32.3",
    "rsa==4.9.1",
    "six==1.17.0",
    "slowapi==0.1.9",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "starlette==0.46.2",
    "stripe>=12.3.0",
    "supabase>=2.16.0",
    "tinycss2>=1.4.0",
    "tomli==2.2.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "uvicorn==0.34.3",
]
